Return to `updateNavigationSpan`
